requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
